Reorganize shims and types
