+++ packages/compile-common/src/shims/index.ts
@@ -1,0 +1,1 @@
+export * as newToLegacy from "./newToLegacy";
